move blind_select logic to its own file
